refactor: move the capability registry into cockpit-registry and example tooling into scripts/examples
