+++ .github/workflows/deploy-langgraph.yml
@@ -7,7 +7,7 @@ on:
       - 'cockpit/langgraph/**/python/**'
       - 'cockpit/deep-agents/**/python/**'
       - 'examples/chat/python/**'
-      - 'apps/cockpit/scripts/capability-registry.ts'
+      - 'libs/cockpit-registry/src/lib/capability-registry.ts'
       - 'scripts/generate-shared-deployment-config.ts'
       - 'deployments/shared-dev/langgraph.json'
   workflow_dispatch:
