+++ apps/website/content/docs/runtimes/mastra/quickstart.mdx
@@ -52,7 +52,7 @@ Export only the variables you need. A stray `AG_UI_INTERNAL_TOKEN` that disagree
 <Step title="Start the Angular app in a second terminal">
 
 ```bash
-npx tsx apps/cockpit/scripts/serve-example.ts --capability=rt-mastra
+npx tsx scripts/examples/serve-example.ts --capability=rt-mastra
 ```
 
 Or run the target directly:
