+++ apps/website/content/docs/runtimes/microsoft-agent-framework/quickstart.mdx
@@ -50,7 +50,7 @@ Azure addresses models by the deployment name you chose in your resource, which 
 <Step title="Start both halves">
 
 ```bash
-npx tsx apps/cockpit/scripts/serve-example.ts --capability=rt-maf
+npx tsx scripts/examples/serve-example.ts --capability=rt-maf
 ```
 
 The script runs `uv sync` in `cockpit/runtimes/microsoft-agent-framework/python` on the way, so the first start takes longer than later ones.
