+++ deployments/ag-ui-dev/README.md
@@ -9,7 +9,7 @@ Multi-topic FastAPI app hosting the cockpit `ag-ui/*` runtimes on Railway.
 | `Dockerfile`, `entrypoint.sh`, `railway.json`, `README.md` | hand-written | yes |
 | `server.py`, `requirements.txt`, `deps/` | generated by `scripts/generate-ag-ui-deployment-config.ts` | no |
 
-The generator reads `apps/cockpit/scripts/capability-registry.ts`, filters to
+The generator reads `libs/cockpit-registry/src/lib/capability-registry.ts`, filters to
 `product === 'ag-ui'` and `product === 'runtimes'` entries with a `pythonDir`,
 stages each into `deps/<topic>/`, and writes one framework-appropriate mount
 (`add_langgraph_fastapi_endpoint`, `add_agent_framework_fastapi_endpoint`, or
