+++ scripts/examples/runtime-wiring-audit.ts
